+++ client/src/components/Home.jsx
@@ -6,7 +6,7 @@ import LeftComponent from './LeftComponent/LeftComponent.jsx';
 import RightComponent from './RightComponent/RightComponent.jsx';
 import Footer from './Footer.jsx';
 import CenterComponent from './CenterComponent/CenterComponent.jsx';
-import { searchGoogle, getNYTimesList, getNYTimesCategory} from '../requests/getRequest.js';
+import { searchGoogle, getNYTimesList, getNYTimesCategory} from '../requests';
 import sample from './RightComponent/TopRankingBooks/sample.js';
 
 export default function Home({ authStatus, authenticate, currentUser }) {
@@ -97,6 +97,7 @@ export default function Home({ authStatus, authenticate, currentUser }) {
     //this route can take a page and count and they can be change, max count is 40
   searchGoogle(query)
   .then((res) => {
+    console.log(res)
     setSearchedBooks(res);
   })
   .catch(err => console.err)
